+++ pkgs/unified_analytics/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added optional `isExternal` parameter to `Analytics.fake` to allow overriding
   the build configuration during testing.
+- Dropped dependency on `package:platform` in favor of `dart:io`.
 
 ## 8.0.18